requirements.txt
pyyaml
pandas
matplotlib
